requirements.txt
pillow~=11.2.1
tensorflow~=2.19.0
numpy~=2.1.3
matplotlib~=3.10.3
gdown~=5.2.0
pytesseract~=0.3.13
tqdm~=4.67.1
doctr~=1.9.0
keras~=3.10.0
huggingface-hub~=0.31.4
protobuf~=3.20.3
SQLAlchemy~=2.0.41
dotenv~=0.9.9
python-dotenv~=1.1.0
Flask~=3.1.1